refactor: extract content from zns.tsx and extract to app.tsx
