--- src/pages/ZNS/index.ts
@@ -1,1 +1,0 @@
-export * from './ZNS';
